# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/organizations/{organizationId}/campaigns": {
      "post": {
        "operationId": "CampaignController_create",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "objective",
                  "channel",
                  "startsAt",
                  "endsAt",
                  "budgetPlannedMinor",
                  "currency"
                ],
                "additionalProperties": false,
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200,
                    "pattern": "\\S"
                  },
                  "objective": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500,
                    "pattern": "\\S"
                  },
                  "channel": {
                    "type": "string",
                    "enum": [
                      "META",
                      "GOOGLE",
                      "SNAPCHAT",
                      "TIKTOK",
                      "X",
                      "LINKEDIN",
                      "PRINT",
                      "OUTDOOR",
                      "REFERRAL",
                      "OTHER"
                    ]
                  },
                  "startsAt": {
                    "type": "string",
                    "format": "date-time",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$"
                  },
                  "endsAt": {
                    "type": "string",
                    "format": "date-time",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$"
                  },
                  "budgetPlannedMinor": {
                    "type": "string",
                    "pattern": "^[1-9]\\d*$"
                  },
                  "currency": { "type": "string", "pattern": "^[A-Z]{3}$" },
                  "utmSource": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100,
                    "pattern": "\\S"
                  },
                  "utmMedium": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100,
                    "pattern": "\\S"
                  },
                  "utmCampaign": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100,
                    "pattern": "\\S"
                  },
                  "utmContent": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100,
                    "pattern": "\\S"
                  },
                  "utmTerm": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100,
                    "pattern": "\\S"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "" },
          "400": { "description": "" },
          "401": { "description": "" },
          "403": { "description": "" }
        },
        "summary": "",
        "tags": ["Campaigns"]
      },
      "get": {
        "operationId": "CampaignController_list",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 512,
              "pattern": "^[A-Za-z0-9_-]+$"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["DRAFT", "ACTIVE", "COMPLETED", "CANCELLED"]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["items"],
                  "additionalProperties": false,
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "id",
                          "name",
                          "objective",
                          "channel",
                          "status",
                          "startsAt",
                          "endsAt",
                          "budgetPlannedMinor",
                          "currency",
                          "touchCount",
                          "createdAt"
                        ],
                        "additionalProperties": false,
                        "properties": {
                          "id": { "type": "string", "format": "uuid" },
                          "name": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 200
                          },
                          "objective": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 500
                          },
                          "channel": {
                            "type": "string",
                            "enum": [
                              "META",
                              "GOOGLE",
                              "SNAPCHAT",
                              "TIKTOK",
                              "X",
                              "LINKEDIN",
                              "PRINT",
                              "OUTDOOR",
                              "REFERRAL",
                              "OTHER"
                            ]
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "DRAFT",
                              "ACTIVE",
                              "COMPLETED",
                              "CANCELLED"
                            ]
                          },
                          "startsAt": {
                            "type": "string",
                            "format": "date-time",
                            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$"
                          },
                          "endsAt": {
                            "type": "string",
                            "format": "date-time",
                            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$"
                          },
                          "budgetPlannedMinor": {
                            "type": "string",
                            "pattern": "^[1-9]\\d*$"
                          },
                          "currency": {
                            "type": "string",
                            "pattern": "^[A-Z]{3}$"
                          },
                          "budgetActualMinor": {
                            "type": "string",
                            "pattern": "^[1-9]\\d*$"
                          },
                          "touchCount": { "type": "integer", "minimum": 0 },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time",
                            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$"
                          }
                        }
                      }
                    },
                    "nextCursor": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 512,
                      "pattern": "^[A-Za-z0-9_-]+$"
                    }
                  }
                }
              }
            },
            "description": ""
          },
          "400": { "description": "" },
          "401": { "description": "" },
          "403": { "description": "" }
        },
        "summary": "",
        "tags": ["Campaigns"]
      }
    },
...
    "/organizations/{organizationId}/campaigns/{campaignId}/transition": {
      "post": {
        "operationId": "CampaignController_transition",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "campaignId",
            "required": true,
            "in": "path",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["toStatus"],
                "additionalProperties": false,
                "properties": {
                  "toStatus": {
                    "type": "string",
                    "enum": ["ACTIVE", "COMPLETED", "CANCELLED"]
                  },
                  "reason": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500,
                    "pattern": "\\S"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "" },
          "400": { "description": "" },
          "401": { "description": "" },
          "403": { "description": "" },
          "404": { "description": "" },
          "409": { "description": "" }
        },
        "summary": "",
        "tags": ["Campaigns"]
      }
    },
    "/organizations/{organizationId}/campaigns/{campaignId}/budget-corrections": {
      "post": {
        "operationId": "CampaignController_correctBudget",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "campaignId",
            "required": true,
            "in": "path",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["correctedMinor", "reason"],
                "additionalProperties": false,
                "properties": {
                  "correctedMinor": {
                    "type": "string",
                    "pattern": "^[1-9]\\d*$"
                  },
                  "reason": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500,
                    "pattern": "\\S"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "" },
          "400": { "description": "" },
          "401": { "description": "" },
          "403": { "description": "" },
          "404": { "description": "" },
          "409": { "description": "" }
        },
        "summary": "",
        "tags": ["Campaigns"]
      }
    },
...
    "/organizations/{organizationId}/leads/{leadId}/attribution-corrections": {
      "post": {
        "operationId": "CampaignController_correctAttribution",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "leadId",
            "required": true,
            "in": "path",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["reason"],
                "additionalProperties": false,
                "properties": {
                  "correctedCampaignId": { "type": "string", "format": "uuid" },
                  "reason": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500,
                    "pattern": "\\S"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "" },
          "400": { "description": "" },
          "401": { "description": "" },
          "403": { "description": "" },
          "404": { "description": "" }
        },
        "summary": "",
        "tags": ["Campaigns"]
      }
    }
  },
  "info": {
    "title": "EstateFlow API",
    "description": "",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
...
}